# backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI-TestHub API
  version: 1.0.0
  description: |
    Versioned REST API for AI-TestHub. Authenticate with `Authorization: Bearer <session JWT or ath_ API token>`.
    API Tokens are additionally limited by level (`read`, `write`, `admin`) and project scope.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /cases:
    get: { summary: List cases, parameters: [{ name: projectId, in: query, schema: { type: string } }], responses: { '200': { description: Case collection } } }
...
  /plans:
    get: { summary: List test plans, parameters: [{ name: projectId, in: query, schema: { type: string } }], responses: { '200': { description: Plan collection } } }
  /defects:
    get: { summary: List defects, parameters: [{ name: projectId, in: query, schema: { type: string } }, { name: status, in: query, schema: { type: string } }], responses: { '200': { description: Defect collection } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }